fix(assistive): address #269 code review findings
Address confirmed code-review findings on the prior commit's fix:

- Move createCharacterCountLength (now createCharacterCountLengthSignal, a
  clearer name matching this dir's create*Signal convention) out of the
  public /headless barrel into packages/toolkit/core/utilities/, the
  documented internal cross-entry plumbing point assistive already uses.
  Revert the headless/src/index.ts barrel addition; both headless/utilities.ts
  and assistive/character-count.ts now import the helper from
  '@ngx-signal-forms/toolkit/core'.
- Drop the malformed `@default component 'createCharacterCount'` JSDoc tag —
  `component` is now a required parameter (both real callers already pass
  one explicitly), documented per core's `@param name - description` style
  plus a `@remarks Does not require an injection context` note.
- Pass `component: 'NgxFormFieldCharacterCount'` to the `createCharacterCount`
  call in the with-maxLength path, so its unsupported-value warning matches
  the no-maxLength fallback's component name instead of falling back to the
  factory's own default ('createCharacterCount').
- Strengthen the "no raw value" spec to assert every raw console.warn
  argument is typeof 'string', not just that its String()-coerced form
  omits certain substrings — the prior version would have let a leaked
  object argument slip through as '[object Object]'.

Refs #269
