+++ packages/toolkit/headless/src/index.ts
@@ -105,7 +105,6 @@ export {
 // Utility functions
 export {
   createCharacterCount,
-  createCharacterCountLength,
   createErrorState,
   createFieldStateFlags,
   createUniqueId,
